+++ test/db/isoexpress/morning_report_test.dart
@@ -25,6 +25,13 @@ Future<void> tests() async {
       var columns = data.first.toJson().keys.toList();
       expect(columns.length, MorningReport.colnames.length);
     });
+    test('read file for 2020-09-28', () async {
+      var file = archive.getFilename(Date.utc(2020, 9, 28));
+      var data = archive.processFile(file);
+      expect(data.length, 2);
+      var columns = data.first.toJson().keys.toList();
+      expect(columns.length, MorningReport.colnames.length);
+    });
     test('read file for 2023-06-16', () async {
       var file = archive.getFilename(Date.utc(2023, 6, 16));
       var data = archive.processFile(file);